requirements.txt
tvdatafeed
python-telegram-bot==20.7
python-dotenv
pandas